fix: delete student

+++ src/interactors/studentInteractor.ts
@@ -1,12 +1,13 @@
-import { getStudentByGraduation as fetchStudentsByGraduation } from './../repositories/studentRepository';
 import * as StudentService from '../services/studentService';
 import * as UserService from '../services/userService';
 import * as UserProfileService from '../services/userProfileService';
 import * as UserRoleService from '../services/userRoleService';
+import * as GraduationService from '../services/graduationService';
 import createUserRequest from '../dtos/createUserRequest';
 import { NotFoundError } from '../errors/notFoundError';
 import { HttpError } from '../errors/httpError';
-import createStudentRequest from 'src/dtos/createStudentRequest';
+import createStudentRequest from '../dtos/createStudentRequest';
+import { ConflictError } from '../errors/conflictError';
 
 const studentRole = 1;
 
@@ -41,7 +42,7 @@ export const createStudent = async (studentData: createStudentRequest) => {
     if (existingUserWithCode) {
       throw new HttpError(409, 'Ya existe un estudiante con este código.');
     }
-    
+
     const { is_scholarship, ...userData } = studentData;
     const newStudent = await UserService.createUser(userData);
 
@@ -57,7 +58,7 @@ export const createStudent = async (studentData: createStudentRequest) => {
 
     await StudentService.createStudent({
       ...studentData,
-      id: newStudent.id
+      id: newStudent.id,
     });
 
     return newStudent;
@@ -77,10 +78,16 @@ export const deleteStudent = async (studentId: number) => {
       throw new NotFoundError('Student not found');
     }
 
+    const process = await GraduationService.getProcessByStudentId(studentId);
+
+    if (!process) {
+      throw new ConflictError('Cannot delete student: Graduation process related exists');
+    }
+
     await UserProfileService.deleteUser(studentId);
   } catch (error) {
     console.error('Error deleting student:', error);
-    throw new Error((error as Error).message);
+    throw error;
   }
 };
